Extract cache logic to separate files
