+++ crates/apps/worker/src/handlers/common/mod.rs
@@ -1,4 +1,5 @@
 pub mod cache;
 pub mod clustering;
+pub mod daily_cards;
 pub mod remote_user;
 pub mod utils;
