+++ crates/apps/worker/src/handlers/ingest_llm.rs
@@ -1,6 +1,5 @@
 use crate::context::WorkerContext;
 use crate::handlers::JobResult;
-use crate::handlers::common::cache::{get_llm_cache, write_llm_cache};
 use crate::handlers::common::utils::get_images_to_analyze;
 use crate::jobs::management::is_job_cancelled;
 use color_eyre::eyre::{Result, eyre};
@@ -11,6 +10,7 @@ use common_services::database::visual_analysis_store::VisualAnalysisStore;
 use common_types::ml_analysis::MLChatAnalysis;
 use std::path::Path;
 use tracing::{debug, info};
+use crate::handlers::common::cache::llm_cache::{get_llm_cache, write_llm_cache};
 
 /// Handles the llm analysis of a given job.
 ///
