+++ crates/libs/generate_thumbnails/src/utils.rs
